Module nvd_cvss::v2

Common Vulnerability Scoring System Version 2.0

Currently, IT management must identify and assess vulnerabilities across many disparate hardware and software platforms. They need to prioritize these vulnerabilities and remediate those that pose the greatest risk. But when there are so many to fix, with each being scored using different scales, how can IT managers convert this mountain of vulnerability data into actionable information? The Common Vulnerability Scoring System (CVSS) is an open framework that addresses this issue. It offers the following benefits:

Standardized Vulnerability Scores:

When an organization normalizes vulnerability scores across all of its software and hardware platforms, it can leverage a single vulnerability management policy. This policy may be similar to a service level agreement (SLA) that states how quickly a particular vulnerability must be validated and remediated.

Open Framework:

Users can be confused when a vulnerability is assigned an arbitrary score. “Which properties gave it that score? How does it differ from the one released yesterday?” With CVSS, anyone can see the individual characteristics used to derive a score.

Prioritized Risk:

When the environmental score is computed, the vulnerability now becomes contextual. That is, vulnerability scores are now representative of the actual risk to an organization. Users know how important a given vulnerability is in relation to other vulnerabilities.

Modules

access_complexity

2.1.2. Access Complexity (AC)

access_vector

2.1.1. Access Vector (AV)

authentication

2.1.3. Authentication (Au)

impact_metrics

CIA Impact Metric

Structs

CVSS

ImpactMetricV2

cvss v2