Module nvd_cvss::v3

Common Vulnerability Scoring System version 3.1

CVSS Version 3.1 Release

This page updates with each release of the CVSS standard. It is currently CVSS version 3.1, released in June 2019. If you wish to use a specific version of the Specification Document, use:

https://www.first.org/cvss/v3.1/specification-document for CVSS version 3.1

https://www.first.org/cvss/v3.0/specification-document for CVSS version 3.0

Modules

attack_complexity

2.1.2. Attack Complexity (AC)

attack_vector

2.1.1. Attack Vector (AV)

impact_metrics

2.3. Impact Metrics

privileges_required

2.1.3. Privileges Required (PR)

scope

2.2. Scope (S)

user_interaction

2.1.4. User Interaction (UI)

Structs

CVSS

The Common Vulnerability Scoring System (CVSS) captures the principal technical characteristics of software, hardware and firmware vulnerabilities. Its outputs include numerical scores indicating the severity of a vulnerability relative to other vulnerabilities.

CVSSBuilder

ExploitAbility

2.1. Exploitability Metrics

ImpactMetricV3

cvss v3